# backend/internal/delivery/http/openapi.yaml
openapi: 3.0.3
info:
  title: Platform RT API
  description: |
    OpenAPI 3.0 specification for the Sitransparan RT/RW backend.

    Autentikasi: `Authorization: Bearer <jwt>` (HS256, valid 24 jam).
    Role: `superadmin`, `admin_rt`, `resident`.
    Konvensi akses pada deskripsi endpoint:
    - PUBLIC: tanpa autentikasi
    - AUTH: semua role terautentikasi
    - ADMIN: superadmin + admin_rt
    - SUPERADMIN: superadmin saja

    Tenant context hanya diambil dari JWT claims yang terverifikasi (header
    X-Tenant-ID, query param, dan subdomain TIDAK dipercaya). Endpoint publik
    `/api/v1/t/{slug}/...` me-resolve tenant dari slug di path.
  version: 2.0.0
servers:
  - url: http://localhost:8081
    description: Local server (Port 8081)
  - url: /
    description: Current host relative path
paths:
...
  /api/v1/reactions:
    get:
      summary: Ringkasan reaksi satu target (AUTH). Query target_type & target_id.
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: target_type, in: query, required: true, schema: { type: string, enum: [announcement, event, meeting] } }
        - { name: target_id, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: '{ counts: {support|like|applause: n}, mine, total }'
    post:
      summary: Beri/ubah reaksi — identitas dari JWT, 1 user 1 reaksi per target (AUTH)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_type, target_id, reaction]
              properties:
                target_type: { type: string, enum: [announcement, event, meeting] }
                target_id: { type: string, format: uuid }
                reaction: { type: string, enum: [support, like, applause] }
      responses:
        '200':
          description: Summary reaksi terkini
        '401':
          description: Login diperlukan
    delete:
      summary: Tarik reaksi (AUTH). Query target_type & target_id.
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Summary reaksi terkini
  /api/v1/polls:
    get:
      summary: List polling berstatus open (AUTH, menyertakan my_vote)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: '{ data: [{ id, question, options, status, votes, total_votes, my_vote }] }'
    post:
      summary: Buat polling (ADMIN). 2..6 opsi.
      security: [{ BearerAuth: [] }]
      responses:
        '201':
          description: Polling dibuat
        '403':
          description: Bukan admin
  /api/v1/polls/{id}:
    get:
      summary: Detail polling + hasil agregat (AUTH, menyertakan my_vote)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Poll
        '404':
          description: Tidak ditemukan
    delete:
      summary: Tutup polling (ADMIN)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Polling ditutup
  /api/v1/polls/{id}/vote:
    post:
      summary: Beri/ubah suara — 1 warga 1 suara (AUTH, unique constraint)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [option_index]
              properties:
                option_index: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Poll dengan hasil terkini
        '400':
          description: Polling tutup / option_index di luar jangkauan
  /api/v1/t/{slug}/polls:
    get:
      summary: Daftar polling terbuka untuk portal publik (PUBLIC, hasil agregat, tanpa voter identity)
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: '{ data: [{ id, question, options, status, votes, total_votes }] }'
        '404':
          description: Tenant tidak ditemukan
  /api/v1/t/{slug}/polls/{id}:
    get:
      summary: Hasil agregat polling untuk portal publik (PUBLIC, tanpa my_vote)
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: '{ id, question, options, status, votes, total_votes }'
        '404':
          description: Tenant/polling tidak ditemukan
  /api/v1/t/{slug}/events:
    get:
      summary: List agenda mendatang untuk portal publik (PUBLIC, 20 terbaru, sorted)
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: '{ data: [{ id, title, description, event_date, location, status }] }'
        '404':
          description: Tenant tidak ditemukan
    post:
      summary: Catat KPI portal — feed_view / share_opened (PUBLIC, rate-limited)
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_type]
              properties:
                event_type: { type: string, enum: [feed_view, share_opened] }
                target_id: { type: string, format: uuid }
      responses:
        '204':
          description: Tercatat
        '400':
          description: event_type tidak valid

  # Upload guard (semua endpoint upload): maksimal 5 MB per file,
  # hanya JPG/PNG/WebP/PDF. Melebihi -> 400.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...